Implement test for update posts with and without authentication, add an error management when a record is not found, create a concern called secured.rb with authenticate_user! method and delete repeated and old tests for create and update posts
