+++ app/controllers/posts_controller.rb
@@ -1,4 +1,5 @@
 class PostsController < ApplicationController
+  include Secured
   before_action :authenticate_user!, only: [ :create, :update ]
 
   rescue_from Exception do |e|
@@ -7,6 +8,10 @@ class PostsController < ApplicationController
 
   rescue_from ActiveRecord::RecordInvalid do |e|
     render json: { error: e.message }, status: :unprocessable_entity
+  end
+
+  rescue_from ActiveRecord::RecordNotFound do |e|
+    render json: { error: e.message }, status: :not_found
   end
 
   # GET /posts
@@ -50,19 +55,5 @@ class PostsController < ApplicationController
 
   def update_params
     params.require(:post).permit(:title, :content, :published)
-  end
-
-  def authenticate_user!
-    token_regex = /Bearer (\w+)/
-    headers = request.headers
-
-    if headers["Authorization"].present? && headers["Authorization"].match(token_regex)
-      token = headers["Authorization"].match(token_regex)[1]
-      if (Current.user = User.find_by_auth_token(token))
-        return
-      end
-    end
-
-    render json: { error: "Unauthorized" }, status: :unauthorized
   end
 end
