+++ spec/requests/posts_spec.rb
@@ -37,7 +37,7 @@ RSpec.describe "Posts endpoints", type: :request do
   end
 
   describe "GET /posts/{id}" do
-    let!(:post) { create(:post) }
+    let!(:post) { create(:post, published: true) }
 
     it "should return a post" do
       get "/posts/#{post.id}"
@@ -51,83 +51,6 @@ RSpec.describe "Posts endpoints", type: :request do
       expect(payload["author"]["email"]).to eq(post.user.email)
       expect(payload["author"]["id"]).to eq(post.user.id)
       expect(response).to have_http_status(200)
-    end
-  end
-
-  describe "POST /posts" do
-    let(:user) { create(:user) }
-
-    it "should create a post" do
-      req_payload = {
-        post: {
-          title: "título",
-          content: "content",
-          published: false,
-          user_id: user.id
-        }
-      }
-
-      # POST HTTP
-      post "/posts", params: req_payload
-      payload = JSON.parse(response.body)
-      expect(payload).to_not be_empty
-      expect(payload["id"]).to_not be_nil
-      expect(response).to have_http_status(:created)
-    end
-
-    it "should return error message on invalid post" do
-      req_payload = {
-        post: {
-          content: "content",
-          published: false,
-          user_id: user.id
-        }
-      }
-
-      # POST HTTP
-      post "/posts", params: req_payload
-      payload = JSON.parse(response.body)
-      expect(payload).to_not be_empty
-      expect(payload["error"]).to_not be_empty
-      expect(response).to have_http_status(:unprocessable_entity)
-    end
-  end
-
-  describe "PUT /posts/{id}" do
-    let(:article) { create(:post) }
-
-    it "should update a post" do
-      req_payload = {
-        post: {
-          title: "título",
-          content: "content",
-          published: true
-        }
-      }
-
-      # PUT HTTP
-      put "/posts/#{article.id}", params: req_payload
-      payload = JSON.parse(response.body)
-      expect(payload).to_not be_empty
-      expect(payload["id"]).to eq(article.id)
-      expect(response).to have_http_status(:ok)
-    end
-
-    it "should return error message on invalid post" do
-      req_payload = {
-        post: {
-          title: nil,
-          content: nil,
-          published: nil
-        }
-      }
-
-      # PUT HTTP
-      put "/posts/#{article.id}", params: req_payload
-      payload = JSON.parse(response.body)
-      expect(payload).to_not be_empty
-      expect(payload["error"]).to_not be_empty
-      expect(response).to have_http_status(:unprocessable_entity)
     end
   end
 end
